Problem: people other than implementors cannot access full MP text
Solution: this might be a small change but it has serious implications
for the coin, so non-techincal readers must easily access the text,
which covers the rationale and motivation for such a change.

To balance the implementors need for succinct information
a new section is added at the top, a TL;DR so to say.