Problem: Link in 1/C4 was dead.
Solution: Remove 'github.com' reference (because Fork+Pull doesn't belong to
or necessitate Github in particular), and replace link with
a personal blog article detailing the process.

I'm not set on that article in particular at all; just wanted to draw
attention to it. Other possible links include but are not limited to:
- https://help.github.com/articles/about-pull-requests/
- https://www.atlassian.com/blog/git/git-branching-and-forking-in-the-enterprise-why-fork